D14 P2 tests passing

+++ src/year_2019/day14/Day14.java
@@ -10,7 +10,7 @@ public class Day14 {
         return reactionInfo.findLeastRequiredOreForOneFuel();
     }
 
-    public static int part2(String fileName) throws IOException {
+    public static Long part2(String fileName) throws IOException {
         ReactionInfo reactionInfo = new ReactionInfo(fileName);
         return reactionInfo.findFuelYouCanMakeWithATrillionOre();
     }
